requirements/base.txt
asgiref==3.5.0
astroid==2.9.3
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-admin-sortable2==1.0.4
django-cleanup==6.0.0
django-model-utils==4.2.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
Pillow==9.0.1
platformdirs==2.5.1
pylint==2.12.2
pyTelegramBotAPI==4.4.0
requests==2.27.1
singlemodeladmin==0.9
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.1.1
urllib3==1.26.8
uWSGI==2.0.20
wrapt==1.13.3
yapf==0.32.0
